+++ VirtualMachineDemo/LC_3/Instruction/TRAPCommand.cs
@@ -13,7 +13,7 @@ namespace LC_3.Instruction
             bitInfo.AddInfo(nameof(this.InstructionSet), 15, 12);
             bitInfo.AddInfo(nameof(this.Trapverct), 7, 0);
         }
-        public int Trapverct { get; set; }
+        public ushort Trapverct { get; set; }
         public override bool Check(string bin)
         {
             return new string(bin.Take(8).ToArray()) == "11110000";
